M2.1
A set of 10 hypothetical patient records from a large database is presented in Table 3.9. Patients with a diabetes value of 1 have type II diabetes and patients with a diabetes value of 0 do not have type II diabetes. It is anticipated that this data set will be used to predict diabetes based on measurements of age, systolic blood pressure, diastolic blood pressure, and weight.